Upgrading the gerrit-rest-java-client dependency to the latest 0.8.11 release.
The latest release contains several API changes which require modification
of the SonarToGerritPublisher class.  In addition, the gerrit-rest-java-client
previously contained a bug which would always authenticate against the /login
page to obtain a session cookie.  When the sonar-gerrit-plugin uses a Gerrit
HTTP password for authentication, this will cause the LDAP account to become
locked due to invalid authentication attempts.

For details about the authentication bug see:
   https://github.com/uwolfer/gerrit-rest-java-client/issues/46